requirements.txt
torch
torchvision
pandas
opencv-python-headless==4.9.0.80
scikit-learn
streamlit
Pillow
google-generativeai
numpy<2.0